changed: code refactoring
